first book return fix

+++ src/main/java/com/app/libraryproject/dto/bookreturn/BookReturnResponse.java
@@ -1,7 +1,9 @@
 package com.app.libraryproject.dto.bookreturn;
 
+import java.math.BigDecimal;
+
 public record BookReturnResponse(
     Long id,
     boolean isLate,
-    double lateFee
+    BigDecimal lateFee
 ) {}
